move use filepath for file and update deps
