+++ gradle/libs.versions.toml
@@ -1,10 +1,10 @@
 [versions]
-kotlin = "1.7.10"
+kotlin = "1.7.20"
 coroutines = "1.6.4"
-serialization = "1.4.0"
-ktor = "2.1.0"
+serialization = "1.4.1"
+ktor = "2.1.3"
 okio = "3.2.0"
-logback = "1.2.11"
+logback = "1.4.4"
 
 [libraries]
 # Coroutines
@@ -39,7 +39,7 @@ ulid = { group = "com.aallam.ulid", name = "ulid-kotlin", version = "1.1.0" }
 [plugins]
 kotlin-multiplaform = { id = "org.jetbrains.kotlin.multiplatform", version.ref = "kotlin" }
 kotlinx-serialization = { id = "org.jetbrains.kotlin.plugin.serialization", version.ref = "kotlin" }
-kotlinx-binary-validator = { id = "org.jetbrains.kotlinx.binary-compatibility-validator", version = "0.11.0" }
-maven-publish = { id = "com.vanniktech.maven.publish", version = "0.21.0" }
-spotless = { id = "com.diffplug.gradle.spotless", version = "6.9.1" }
+kotlinx-binary-validator = { id = "org.jetbrains.kotlinx.binary-compatibility-validator", version = "0.12.1" }
+maven-publish = { id = "com.vanniktech.maven.publish", version = "0.22.0" }
+spotless = { id = "com.diffplug.gradle.spotless", version = "6.11.0" }
 dokka = { id = "org.jetbrains.dokka", version = "1.7.10" }
